move files around and use shared colors file
