+++ packages/app/src/components/AddToLayoutButton/styles/DropDownButton.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../modules/colors';
+import { colors } from 'analytics-shared';
 
 export default {
     arrowDown: {
