+++ packages/app/src/components/DimensionsPanel/Dialogs/DataDimension/DataDimension.js
@@ -8,7 +8,7 @@ import debounce from 'lodash-es/debounce';
 import keyBy from 'lodash-es/keyBy';
 import isEqual from 'lodash-es/isEqual';
 
-import { ItemSelector } from 'analytics-shared-components';
+import { ItemSelector } from 'analytics-shared';
 import DataTypes from './DataTypesSelector';
 import Groups from './Groups';
 import FilterField from '../FilterField';
