+++ packages/app/src/components/DimensionsPanel/Dialogs/DynamicDimension/DynamicDimension.js
@@ -5,7 +5,7 @@ import DialogTitle from '@material-ui/core/DialogTitle';
 import DialogContent from '@material-ui/core/DialogContent';
 import i18n from '@dhis2/d2-i18n';
 import keyBy from 'lodash-es/keyBy';
-import { ItemSelector } from 'analytics-shared-components';
+import { ItemSelector } from 'analytics-shared';
 
 import FilterField from '../FilterField';
