+++ packages/app/src/components/DimensionsPanel/List/styles/DimensionItem.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../../modules/colors';
+import { colors } from 'analytics-shared';
 
 export const styles = {
     text: {
