+++ packages/app/src/components/Layout/DefaultLayout/styles/DefaultAxis.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../../modules/colors';
+import { colors } from 'analytics-shared';
 import defaultLayoutStyle from './DefaultLayout.style';
 import * as layoutStyle from '../../styles/style';
