+++ packages/app/src/components/Snackbar/styles/Snackbar.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../modules/colors';
+import { colors } from 'analytics-shared';
 
 export default theme => ({
     container: {
