+++ packages/app/src/components/TitleBar/styles/TitleBar.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../modules/colors';
+import { colors } from 'analytics-shared';
 
 export default {
     titleBar: {
