+++ packages/app/src/components/Visualization/styles/BlankCanvas.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../modules/colors';
+import { colors } from 'analytics-shared';
 
 const error = {
     fontWeight: 'bold',
