+++ packages/app/src/components/VisualizationTypeSelector/styles/VisualizationTypeSelector.style.js
@@ -1,4 +1,4 @@
-import { colors } from '../../../modules/colors';
+import { colors } from 'analytics-shared';
 
 export default {
     button: {
