+++ packages/shared/src/components/ItemSelector/ItemIcon.js
