+++ packages/shared/src/components/ItemSelector/SelectedItems.js
