+++ packages/shared/src/components/ItemSelector/__tests__/SelectedItems.spec.js
