+++ packages/shared/src/components/ItemSelector/buttons/RemoveDimensionButton.js
