+++ packages/shared/src/components/ItemSelector/styles/Item.style.js
@@ -1,4 +1,5 @@
-import { colors } from '../../colors';
+// import { colors } from 'analytics-shared';
+import { colors } from '../../../modules/colors';
 
 export const styles = {
     highlightedItem: {
